+++ src/mesh/primitive/Spring.cc
@@ -25,8 +25,8 @@ using namespace std;
 shared_ptr<Mesh> Spring::Mesh(float minorRadius,
 							  float majorRadius,
 							  float length,
-							  int slices,
-							  int segments,
+							  unsigned slices,
+							  unsigned segments,
 							  const shared_ptr<Material> material) {
 
 	return make_shared<a3d::Mesh>("Spring",
@@ -45,8 +45,8 @@ shared_ptr<Mesh> Spring::Mesh(float minorRadius,
 Spring::Spring(float minorRadius,
 			   float majorRadius,
 			   float length,
-			   int slices,
-			   int segments):
+			   unsigned slices,
+			   unsigned segments):
 		MeshElement{},
 		_minorRadius{minorRadius},
 		_majorRadius{majorRadius},
@@ -62,7 +62,7 @@ Spring::Spring(float minorRadius,
 	/// @param majorStart Counterclockwise angle around the z-axis relative to the x-axis.
 	/// @param majorSweep Counterclockwise angle arounf the z-axis.
 
-	auto spring = SpringMesh{minorRadius, majorRadius, length/2.0, slices, segments};
+	auto spring = SpringMesh{minorRadius, majorRadius, length/2.0, (int)slices, (int)segments};
 
 	for (const MeshVertex& v : spring.vertices()) {
 		_vertices.push_back({ vec3(v.position[0], v.position[1], v.position[2]),
@@ -93,10 +93,10 @@ float Spring::length() const {
 	return _length;
 }
 
-int Spring::slices() const {
+unsigned Spring::slices() const {
 	return _slices;
 }
 
-int Spring::segments() const {
+unsigned Spring::segments() const {
 	return _segments;
 }
